Bug fix for supervision threads exiting on pre 3.2 agencies

+++ arangod/Agency/Supervision.cpp
@@ -49,7 +49,7 @@ struct HealthRecord {
   std::string status;
   std::string endpoint;
   std::string lastAcked;
-  double version;
+  size_t version;
 
   HealthRecord() : version(0) {}
